Add persistent_async_server fixture
